feat(operator): add namespace to grafanaDashboard (#168)
* 🚀 feat(operator): add namespace to grafanaDashboard
The `namespace` field has been added to the `grafanaDashboard` section of the `values.yaml` file. This allows the Grafana dashboard to be deployed to a specific namespace, which improves organization and separation of concerns.

Signed-off-by: MateSousa <matessousa@outlook.com>

* 🚀 feat(grafana-k8sgpt-dashboard.yaml): add namespace field to metadata
The namespace field is added to the metadata section of the Grafana dashboard YAML file. This allows the dashboard to be created in the specified namespace, which improves organization and management of the dashboard.

Signed-off-by: MateSousa <matessousa@outlook.com>

* 🚀 feat(grafana-k8sgpt-dashboard.yaml): add namespaceSelector to k8sgpt-overview dashboard
The namespaceSelector is added to the k8sgpt-overview dashboard to allow it to be displayed in the namespace specified in the values.yaml file. This improves the flexibility of the dashboard as it can now be displayed in any namespace specified in the values.yaml file.

Signed-off-by: MateSousa <matessousa@outlook.com>

* 🔧 chore(operator): remove unnecessary namespace field from grafanaDashboard in values.yaml
The namespace field in the grafanaDashboard section of the values.yaml file is not needed as it is commented out and has no value. Removing it improves the readability of the file.

Signed-off-by: MateSousa <matessousa@outlook.com>

* 🔧 chore(grafana-k8sgpt-dashboard.yaml): simplify namespace configuration
The namespace configuration for the Grafana dashboard has been simplified by removing the unnecessary namespaceSelector block. Instead, the namespace is directly set using the .Values.grafanaDashboard.namespace value, falling back to the default namespace if not provided. This change improves readability and reduces complexity in the configuration file.

Signed-off-by: MateSousa <matessousa@outlook.com>

---------

Signed-off-by: MateSousa <matessousa@outlook.com>
Co-authored-by: Alex Jones <alexsimonjones@gmail.com>