feat: rework controllers logic & add some simple tests (#498)
* feat: rework controllers logic & add some simple tests

Signed-off-by: Matthis Holleville <matthish29@gmail.com>

* feat: enhance deployment readiness check and version upgrade logic

Improved the logic for checking deployment readiness and handling version upgrades in the Kubernetes controller. This includes a more efficient way of determining if a deployment has ready replicas and streamlining the process for updating the deployment image when a new version is specified. This change aims to reduce unnecessary updates and checks, improving the overall efficiency and reliability of the deployment process.

Signed-off-by: Matthis Holleville <matthish29@gmail.com>

* fix: improve log message formatting in analysis step cleanup

Signed-off-by: Matthis Holleville <matthish29@gmail.com>

* ci: add make as a build tool for testing

Signed-off-by: Matthis Holleville <matthish29@gmail.com>

---------

Signed-off-by: Matthis Holleville <matthish29@gmail.com>
Signed-off-by: Ronald Petty <ronald.petty@rx-m.com>